Created send and listener for msdl for rapid response

+++ epd-common/src/main/java/dk/dma/epd/common/prototype/model/voct/SAROperation.java
@@ -1832,7 +1832,7 @@ public class SAROperation {
             effortAllocationData.setTrackSpacing(trackSpacing);
 
             double groundSpeed = effortAllocationData.getGroundSpeed();
-            int timeSearching = effortAllocationData.getSearchTime();
+            double timeSearching = effortAllocationData.getSearchTime();
 
 //            System.out.println("Track Spacing is: " + trackSpacing);
 //            System.out.println("Ground speed is: " + groundSpeed);
